Move About card component into folder
