+++ client/src/App.js
@@ -1,7 +1,7 @@
 import "./App.css";
 import HomeCardConnector from "./components/HomeCard/HomeCardConnector";
 import Navbar from "./components/Navbar";
-import AboutCard from "./components/AboutCard";
+import AboutCard from "./components/AboutCard/AboutCard";
 import { Route, Routes } from "react-router-dom";
 
 function App() {
